created baseActivity class to fix data loading issues
