+++ app/src/main/AndroidManifest.xml
@@ -26,6 +26,9 @@
             android:label="@string/app_name"
             android:theme="@style/AppTheme">
         </activity>
+        <activity
+            android:name=".BaseActivity">
+        </activity>
     </application>
 
 </manifest>
