+++ app/src/main/java/com/nkv/covid/SplashActivity.java
@@ -1,41 +1,20 @@
 package com.nkv.covid;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
-import android.util.Log;
-import android.widget.Toast;
 
-import androidx.appcompat.app.AppCompatActivity;
-
-import static android.content.ContentValues.TAG;
-
-import com.google.gson.Gson;
-import com.nkv.covid.model.CountryRestModel;
-import com.nkv.covid.model.GlobalRestModel;
-
-import java.util.List;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class SplashActivity extends AppCompatActivity {
+public class SplashActivity extends BaseActivity {
     private final int SPLASH_DISPLAY_LENGTH = 3000;
-
-    SharedPreferences mPrefs;
-    SharedPreferences.Editor prefsEditor;
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash);
         try {
-            MainActivity main = new MainActivity();
-//            main.fetchGlobalData(getWindow().getDecorView());
-//            main.fetchCountriesData(getWindow().getDecorView());
+
+            fetchGlobalData();
+            fetchCountriesData();
             new Handler().postDelayed(new Runnable(){
                 @Override
                 public void run() {
@@ -69,26 +48,4 @@ public class SplashActivity extends AppCompatActivity {
         t.start();
     }
 
-//    private void fetchGlobalData(){
-//        RetrofitApiInterface apiService =
-//                RetrofitApiClient.getClient().create(RetrofitApiInterface.class);
-//        Call<GlobalRestModel> call = apiService.getGlobalStats();
-//        call.enqueue(new Callback<GlobalRestModel>() {
-//            @Override
-//            public void onResponse(Call<GlobalRestModel>call, Response<GlobalRestModel> response) {
-//                GlobalRestModel globalData = response.body();
-//                mPrefs = getSharedPreferences("covidData", MODE_PRIVATE);
-//                prefsEditor = mPrefs.edit();
-//                Gson gson = new Gson();
-//                String json = gson.toJson(globalData);
-//                prefsEditor.putString("globalData", json);
-//                prefsEditor.apply();
-//            }
-//
-//            @Override
-//            public void onFailure(Call<GlobalRestModel>call, Throwable t) {
-//                Toast.makeText(SplashActivity.this, "Failed loading data!", Toast.LENGTH_SHORT).show();
-//            }
-//        });
-//    }
 }
