+++ app/src/main/java/com/nkv/covid/ui/global/GlobalFragment.java
@@ -14,6 +14,8 @@ import com.nkv.covid.MainActivity;
 import com.nkv.covid.R;
 import com.nkv.covid.model.GlobalCardModel;
 
+import java.util.Objects;
+
 public class GlobalFragment extends Fragment {
 
     private SwipeRefreshLayout mSwipeRefreshLayout;
@@ -30,7 +32,7 @@ public class GlobalFragment extends Fragment {
         super.onViewCreated(view, savedInstanceState);
 //        String cases, deaths, recovered;
 
-        ((MainActivity) getActivity()).setActionBarTitle(getString(R.string.title_bar_1));
+        ((MainActivity) Objects.requireNonNull(getActivity())).setActionBarTitle(getString(R.string.title_bar_1));
 
         // /You will setup the action bar with pull to refresh layout
         mSwipeRefreshLayout = (SwipeRefreshLayout) view.findViewById(R.id.swipeContainer);
@@ -38,7 +40,7 @@ public class GlobalFragment extends Fragment {
         mSwipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
             @Override
             public void onRefresh() {
-                ((MainActivity) getActivity()).fetchGlobalData();
+                ((MainActivity) Objects.requireNonNull(getActivity())).fetchGlobalData();
                 mSwipeRefreshLayout.setRefreshing(false);
             }
         });
